+++ CHANGELOG.md
@@ -2,7 +2,11 @@ aki-image-view-gtk TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.2.2 (2021-08-19)
+=====
+
 * add: a support of a image path into command arguments
+* add: a file chooser dialog
 
 0.2.1 (2021-08-16)
 =====